move vaultclient to common package
